Add files via upload
Salt-and-pepper noise is a form of noise sometimes seen in images. It is also known as impulse noise. This noise can be caused by sharp and sudden disturbances in the image signal. It presents itself as sparsely occurring white and black pixels.
To remove this noise we use median filtering :

Median filtering is a nonlinear method used to remove noise from images. It is widely used as it is very effective at removing noise while preserving edges. ... The median filter works by moving through the image pixel by pixel, replacing each value with the median value of neighboring pixels.